Dividing into different files frontoffice code
No functionnal changes expected.

All code was in a single file, scrolling through this file to find
something was cumbersome.

Using ES6 import/export and extracting the main functions into different
files helps solve that problem and better understand how the code
articulates. It makes those extracted function into modules.

For now, this patch only extracts the existing main functions. In a
later patch I suggest to also create different folders and divide more
precisely the existing functions.

(Also, `cttm_metas` variable was not declared).
