+++ includes/public/cttm-shortcode.php
@@ -20,7 +20,7 @@ function cttm_shortcode($attr)
 
     wp_enqueue_script('leaflet');
     wp_enqueue_script('leaflet_markercluster');
-    wp_enqueue_script('travelersmap_init');
+    wp_enqueue_script('travelersmap');
 
     //Get post types selected in plugin settings
     $cttm_options = get_option('cttm_options');
@@ -322,8 +322,8 @@ function cttm_shortcode($attr)
     );
 
     //Send Json variables to our javascript file 'travelersmap.js'
-    wp_localize_script('travelersmap_init', 'cttm_options_params', $cttm_options_params);
-    wp_localize_script('travelersmap_init', 'cttm_shortcode_' . $id, ${"cttm_shortcode_$id"});
+    wp_localize_script('travelersmap', 'cttm_options_params', $cttm_options_params);
+    wp_localize_script('travelersmap', 'cttm_shortcode_' . $id, ${"cttm_shortcode_$id"});
     if ($cttm_metas) {
         $cttm_output =   '<div id="' . $containerid . '" class="travelersmap-container" style="z-index: 1; min-height: 10px; min-width:10px; height:' . $height . ';width:' . $width . '; max-width:' . $maxwidth . '; max-height:' . $maxheight . '; position:relative;"><div style="position:absolute; z-index:-1;top: 50%;text-align: center;display: block;left: 50%;transform: translate(-50%,-50%);">Travelers\' Map is loading... <br> If you see this after your page is loaded completely, leafletJS files are missing.</div></div>';
     } else {
